backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
moviepy==1.0.3
numpy==1.26.4
pydub==0.25.1
python-magic-bin==0.4.14
werkzeug==3.0.1
gunicorn==21.2.0
pytesseract==0.3.10
opencv-python==4.9.0.80
Pillow==10.1.0
tensorflow==2.18.0
transformers==4.36.2
torch==2.2.0
torchaudio==2.2.0
openai-whisper==20231117
librosa==0.10.1
scipy==1.12.0
numba==0.59.0
SpeechRecognition==3.10.0
noisereduce==3.0.2
deep-translator==1.11.4
